+++ bfabric/src/bfabric/rest/token_data.py
@@ -5,11 +5,12 @@ from typing import TYPE_CHECKING, Any
 
 import requests
 from pydantic import BaseModel, Field, SecretStr, ConfigDict
+from requests import JSONDecodeError
 
 from bfabric.entities.core.import_entity import import_entity
 
 if TYPE_CHECKING:
-    from bfabric import BfabricClientConfig, Bfabric
+    from bfabric import Bfabric
     from bfabric.entities import Dataset, Instrument, Order, Plate, Project, Resource, Run, Sample, Workunit
 
 
@@ -46,14 +47,22 @@ class TokenData(BaseModel):
         return entity_class.find(self.entity_id, client=client)
 
 
-def get_token_data(client_config: BfabricClientConfig, token: str) -> TokenData:
+def get_raw_token_data(base_url: str, token: str) -> dict[str, Any]:
+    """Returns the raw token data for the provided token."""
+    url = f"{base_url}/rest/token/validate"
+    response = requests.get(url, params={"token": token})
+    if not response.ok:
+        response.raise_for_status()
+    try:
+        return response.json()
+    except JSONDecodeError:
+        raise RuntimeError(f"Get token data failed with message: {response.text!r}")
+
+
+def get_token_data(base_url: str, token: str) -> TokenData:
     """Returns the token data for the provided token.
 
     If the request fails, an exception is raised.
     """
-    url = f"{client_config.base_url}/rest/token/validate"
-    response = requests.get(url, params={"token": token})
-    if not response.ok:
-        response.raise_for_status()
-    parsed = response.json()
+    parsed = get_raw_token_data(base_url=base_url, token=token)
     return TokenData.model_validate(parsed)