+++ bfabric_app_runner/examples/template/src/{{project_name}}/workflow_params.py
@@ -1,0 +1,6 @@
+from pydantic import BaseModel
+
+
+class WorkflowParams(BaseModel):
+    request_failure: bool
+    """If set to True, the workflow will fail with an error message (for testing purposes)."""