+++ bfabric_app_runner/docs/changelog.md
@@ -4,7 +4,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## \[Unreleased\]
 
+### Changed
+
+- `bfabric-cli prepare workunit` does not accept module refs anymore and will resolve the app spec path.
+
 ## \[0.1.2\] - 2025-07-08
+
+### Changed
 
 - `CommandPythonEnv` with `refresh=True` now will create a separate environment to avoid breaking apps which are already using a particular
     Python environment without locking it.