+++ bfabric_app_runner/docs/changelog.md
@@ -9,6 +9,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - `bfabric-cli prepare workunit` does not accept module refs anymore and will resolve the app spec path.
 - `bfabric-cli prepare workunit --force-app-version` to force a specific app version for the workunit.
 
+### Added
+
+- `copier` based template/demo application for development and end-to-end testing of bfabric-app-runner.
+
 ## \[0.1.2\] - 2025-07-08
 
 ### Changed